Improved declarations and moved deprecates to Oldie.
