+++ ts/Core/Globals.ts
@@ -17,7 +17,6 @@
  * */
 
 import type GlobalsLike from './GlobalsLike';
-import type Options from './Options';
 
 /* *
  *
@@ -26,14 +25,19 @@ import type Options from './Options';
  * */
 
 declare global {
+
     type AnyRecord = Record<string, any>;
+
     type DeepPartial<T> = {
         [P in keyof T]?: (T[P]|DeepPartial<T[P]>);
     }
+
     type DeepRecord<K extends keyof any, T> = {
         [P in K]: (T|DeepRecord<K, T>);
     }
+
     type ExtractArrayType<T> = T extends (infer U)[] ? U : never;
+
     interface CallableFunction {
         apply<TScope, TArguments extends Array<unknown>, TReturn>(
             this: (this: TScope, ...args: TArguments) => TReturn,
@@ -41,6 +45,7 @@ declare global {
             args?: (TArguments|IArguments)
         ): TReturn;
     }
+
     interface Element {
         /**
          * @private
@@ -57,12 +62,15 @@ declare global {
             value: (boolean|number|string)
         ): void;
     }
+
     interface HTMLElement {
         parentNode: HTMLElement;
     }
+
     interface Math {
         easeInOutSine(pos: number): number;
     }
+
     interface ObjectConstructor {
         /**
          * Sets the prototype of a specified object o to object proto or null.
@@ -72,6 +80,7 @@ declare global {
          */
         setPrototypeOf?<T>(o: T, proto: object | null): T;
     }
+
     interface SVGElement {
         /**
          * @private
@@ -80,15 +89,18 @@ declare global {
         cutHeight?: number;
         parentNode: SVGElement;
     }
+
     interface TouchList {
         changedTouches: Array<Touch>;
     }
+
     /**
      * @private
      * @deprecated
      * @todo: Rename UMD argument `win` to `window`
      */
     const win: Window|undefined;
+
 }
 
 /* *
